// src/app/interfaces/UserRenewGoogle.interface.ts
// Generated by https://quicktype.io

import { UsuarioModel } from "../models/usuario.model";
import { UsuarioDB } from "./UsuarioDB.inetrface";

export interface UserRenewGoogle {
    ok:        boolean;
    uid:       string;
    token:     string;
    usuarioDB: UsuarioModel;
    menu: string;
}